src: integrate Nan into node
Add Nan to (node_root_dir)/include/node directory and its header files will
be included in the node-gyp build. When creating an addon, include <nan.h> in
the source file and there is no need to install Nan  explicitly.

The silly path (include/node) is used to keep node-gyp working with --nodedir.